feat: lock individual response navbar, fix mobile view for individual responses, add verified label to verified question labels
