+++ frontend/src/features/admin-form/responses/IndividualResponsePage/IndividualResponsePage.tsx
@@ -1,22 +1,9 @@
 import { memo, useCallback, useMemo } from 'react'
-import {
-  BiChevronLeft,
-  BiChevronRight,
-  BiDownload,
-  BiLeftArrowAlt,
-} from 'react-icons/bi'
-import {
-  Link as ReactLink,
-  useLocation,
-  useNavigate,
-  useParams,
-} from 'react-router-dom'
+import { BiDownload } from 'react-icons/bi'
+import { useParams } from 'react-router-dom'
 import {
   Box,
-  ButtonGroup,
-  Grid,
-  Icon,
-  Link,
+  Flex,
   Skeleton,
   Stack,
   StackDivider,
@@ -25,16 +12,15 @@ import {
 import simplur from 'simplur'
 
 import Button from '~components/Button'
-import IconButton from '~components/IconButton'
 import Spinner from '~components/Spinner'
 
 import {
   SecretKeyVerification,
   useStorageResponsesContext,
 } from '../ResponsesPage/storage'
-import { useUnlockedResponses } from '../ResponsesPage/storage/UnlockedResponses/UnlockedResponsesProvider'
 
 import { DecryptedRow } from './DecryptedRow'
+import { IndividualResponseNavbar } from './IndividualResponseNavbar'
 import { useMutateDownloadAttachments } from './mutations'
 import { useIndividualSubmission } from './queries'
 
@@ -60,86 +46,11 @@ const LoadingDecryption = memo(() => {
 })
 
 export const IndividualResponsePage = (): JSX.Element => {
-  const { state } = useLocation()
-  const navigate = useNavigate()
   const { submissionId } = useParams()
   if (!submissionId) throw new Error('Missing submissionId')
 
-  const currentRespondentNumber = useMemo(() => {
-    return (state as { respondentNumber?: number })?.respondentNumber
-  }, [state])
-
   const { secretKey } = useStorageResponsesContext()
-  const {
-    lastNavPage,
-    lastNavSubmissionId,
-    getNextSubmissionId,
-    getPreviousSubmissionId,
-    onNavNextSubmissionId,
-    onNavPreviousSubmissionId,
-    isAnyFetching,
-  } = useUnlockedResponses()
   const { data, isLoading, isError } = useIndividualSubmission()
-
-  const nextSubmissionId = useMemo(
-    () => getNextSubmissionId(submissionId),
-    [getNextSubmissionId, submissionId],
-  )
-  const prevSubmissionId = useMemo(
-    () => getPreviousSubmissionId(submissionId),
-    [getPreviousSubmissionId, submissionId],
-  )
-
-  const handleNavigateNext = useCallback(() => {
-    if (!nextSubmissionId) return
-    navigate(`../${nextSubmissionId}`, {
-      state: {
-        respondentNumber: currentRespondentNumber
-          ? currentRespondentNumber - 1
-          : undefined,
-      },
-    })
-    onNavNextSubmissionId(submissionId)
-  }, [
-    currentRespondentNumber,
-    navigate,
-    nextSubmissionId,
-    onNavNextSubmissionId,
-    submissionId,
-  ])
-
-  const handleNavigatePrev = useCallback(() => {
-    if (!prevSubmissionId) return
-    navigate(`../${prevSubmissionId}`, {
-      state: {
-        respondentNumber: currentRespondentNumber
-          ? currentRespondentNumber + 1
-          : undefined,
-      },
-    })
-    onNavPreviousSubmissionId(submissionId)
-  }, [
-    currentRespondentNumber,
-    navigate,
-    onNavPreviousSubmissionId,
-    prevSubmissionId,
-    submissionId,
-  ])
-
-  const backLink = useMemo(() => {
-    if (!lastNavPage && !lastNavSubmissionId) {
-      return `..`
-    }
-
-    const searchParams = new URLSearchParams()
-    if (lastNavPage) {
-      searchParams.set('page', lastNavPage.toString())
-    }
-    if (lastNavSubmissionId) {
-      searchParams.set('submissionId', lastNavSubmissionId)
-    }
-    return `..?${searchParams}`
-  }, [lastNavPage, lastNavSubmissionId])
 
   const attachmentDownloadUrls = useMemo(() => {
     const attachmentDownloadUrls = new Map()
@@ -172,43 +83,13 @@ export const IndividualResponsePage = (): JSX.Element => {
   if (!secretKey) return <SecretKeyVerification />
 
   return (
-    <Grid
-      px={{ base: '1.5rem', md: '1.75rem', lg: '2rem' }}
-      columnGap={{ base: '0.5rem', lg: '1rem' }}
-      templateColumns={{ base: 'repeat(4, 1fr)', md: 'repeat(12, 1fr)' }}
-    >
-      <Stack spacing="2.5rem" gridColumn="2 / -2">
-        <Stack direction="row" justify="space-between" align="center">
-          <Link
-            display="inline-flex"
-            as={ReactLink}
-            variant="standalone"
-            to={backLink}
-          >
-            <Icon as={BiLeftArrowAlt} fontSize="1.5rem" mr="0.5rem" />
-            Back to list
-          </Link>
-          <Skeleton isLoaded={!isLoading}>
-            <Text textStyle="h2" as="h2">
-              Respondent
-              {currentRespondentNumber ? ` #${currentRespondentNumber}` : ''}
-            </Text>
-          </Skeleton>
-          <ButtonGroup>
-            <IconButton
-              isDisabled={!prevSubmissionId || isAnyFetching}
-              onClick={handleNavigatePrev}
-              icon={<BiChevronLeft />}
-              aria-label="Previous submission"
-            />
-            <IconButton
-              isDisabled={!nextSubmissionId || isAnyFetching}
-              onClick={handleNavigateNext}
-              icon={<BiChevronRight />}
-              aria-label="Next submission"
-            />
-          </ButtonGroup>
-        </Stack>
+    <Flex flexDir="column" marginTop={{ base: '-1.5rem', md: '-3rem' }}>
+      <IndividualResponseNavbar />
+
+      <Stack
+        px={{ md: '1.75rem', lg: '2rem' }}
+        spacing={{ base: '1.5rem', md: '2.5rem' }}
+      >
         <Stack
           bg="primary.100"
           p="1.5rem"
@@ -216,27 +97,39 @@ export const IndividualResponsePage = (): JSX.Element => {
             fontFeatureSettings: "'tnum' on, 'lnum' on, 'zero' on, 'cv05' on",
           }}
         >
-          <Text display="inline-flex">
+          <Stack
+            spacing={{ base: '0', md: '0.5rem' }}
+            direction={{ base: 'column', md: 'row' }}
+          >
             <Text as="span" textStyle="subhead-1">
               Response ID:
             </Text>
-            &nbsp;{submissionId}
-          </Text>
-          <Box display="inline-flex">
+            <Text>{submissionId}</Text>
+          </Stack>
+          <Stack
+            spacing={{ base: '0', md: '0.5rem' }}
+            direction={{ base: 'column', md: 'row' }}
+          >
             <Text as="span" textStyle="subhead-1">
               Timestamp:
             </Text>
             <Skeleton isLoaded={!isLoading && !isError}>
-              &nbsp;{data?.submissionTime ?? 'Loading...'}
+              {data?.submissionTime ?? 'Loading...'}
             </Skeleton>
-          </Box>
+          </Stack>
           {attachmentDownloadUrls.size > 0 && (
-            <Box alignItems="center" display="inline-flex">
-              <Text as="span" textStyle="subhead-1">
+            <Stack
+              spacing={{ base: '0', md: '0.5rem' }}
+              direction={{ base: 'column', md: 'row' }}
+            >
+              <Text
+                as="span"
+                textStyle="subhead-1"
+                py={{ base: '0', md: '0.25rem' }}
+              >
                 Attachments:
               </Text>
               <Skeleton isLoaded={!isLoading && !isError}>
-                &nbsp;
                 <Button
                   variant="link"
                   isDisabled={downloadAttachmentsAsZipMutation.isLoading}
@@ -252,9 +145,10 @@ export const IndividualResponsePage = (): JSX.Element => {
                   {simplur`Download ${attachmentDownloadUrls.size} attachment[|s] as .zip`}
                 </Button>
               </Skeleton>
-            </Box>
+            </Stack>
           )}
         </Stack>
+
         <Stack>
           {isLoading || isError ? (
             <LoadingDecryption />
@@ -268,6 +162,6 @@ export const IndividualResponsePage = (): JSX.Element => {
           )}
         </Stack>
       </Stack>
-    </Grid>
+    </Flex>
   )
 }
